refactor(client) another attempt at drag/drop
Its still not great when dropping a card, drawing a card or
discarding cards. For instance after flicking a card out they
immediately jump before going to the new location.
Discarded cards just disappear instead of animating to the
new location.  This should be simply but its proving to be a
real challenge as the animation needs to respond to changes
in the from/to locations without triggering when they are
changing to the animation itself.